fix/execute-setup-double-code
