+++ pipelex/pipeline/__init__.py
@@ -1,0 +1,7 @@
+from pipelex.pipeline.run.execute import execute_pipeline
+from pipelex.pipeline.run.start import start_pipeline
+
+__all__ = [
+    "execute_pipeline",
+    "start_pipeline",
+]
