+++ tests/e2e/pipelex/pipes/pipe_controller/pipe_sequence/test_pipe_sequence_multiplicity.py
@@ -7,7 +7,7 @@ from pipelex.core.memory.working_memory_factory import WorkingMemoryFactory
 from pipelex.core.stuffs.stuff_factory import StuffFactory
 from pipelex.core.stuffs.text_content import TextContent
 from pipelex.pipe_run.pipe_run_params import PipeRunMode
-from pipelex.pipeline.execute import execute_pipeline
+from pipelex.pipeline import execute_pipeline
 
 
 @pytest.mark.dry_runnable
@@ -33,7 +33,7 @@ class TestPipeSequenceMultiplicity:
 
         # Execute the pipeline
         pipe_output = await execute_pipeline(
-            library_path="tests/integration/pipelex/pipes/controller/pipe_sequence/",
+            library_dirs=["tests/integration/pipelex/pipes/controller/pipe_sequence/"],
             pipe_code="creative_ideation_sequence",
             inputs=working_memory,
             pipe_run_mode=pipe_run_mode,
